smart-crosswalk/backend/ai/requirements.txt
ultralytics>=8.3.0
opencv-python>=4.10.0
numpy>=1.26.0
requests>=2.32.0
cloudinary>=1.41.0
python-dotenv>=1.0.0
